improvements in preview, quicklook

+++ cc/read-file.cc
@@ -231,8 +231,8 @@ void acmacs::file::write(std::string_view aFilename, std::string_view aData, for
 
 // ----------------------------------------------------------------------
 
-acmacs::file::temp::temp(std::string prefix, std::string suffix)
-    : name(make_template(prefix) + suffix), fd(mkstemps(const_cast<char*>(name.c_str()), static_cast<int>(suffix.size())))
+acmacs::file::temp::temp(std::string prefix, std::string suffix, bool autoremove)
+    : name(make_template(prefix) + suffix), autoremove_{autoremove}, fd(mkstemps(const_cast<char*>(name.c_str()), static_cast<int>(suffix.size())))
 {
     if (fd < 0)
         throw std::runtime_error(std::string("Cannot create temporary file using template ") + name + ": " + strerror(errno));
@@ -241,8 +241,8 @@ acmacs::file::temp::temp(std::string prefix, std::string suffix)
 
 // ----------------------------------------------------------------------
 
-acmacs::file::temp::temp(std::string suffix)
-    : acmacs::file::temp::temp("AD.", suffix)
+acmacs::file::temp::temp(std::string suffix, bool autoremove)
+    : acmacs::file::temp::temp("AD.", suffix, autoremove)
 {
 
 } // acmacs::file::temp::temp
@@ -251,7 +251,7 @@ acmacs::file::temp::temp(std::string suffix)
 
 acmacs::file::temp::~temp()
 {
-    if (!name.empty())
+    if (autoremove_ && !name.empty())
         fs::remove(name.c_str());
 
 } // acmacs::file::temp::~temp
